Updated delete method to show if the acitve city being viewed is no longer favorited
