+++ Server/Helpers/JwtTokenGenerator.cs
@@ -1,0 +1,34 @@
+using System.IdentityModel.Tokens.Jwt;
+using System.Security.Claims;
+using System.Text;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.IdentityModel.Tokens;
+using react_weatherapp.Models;
+
+namespace react_weatherapp.Helpers
+{
+    public static class TokenHelper
+    {
+        public static IActionResult GenerateToken(User user, IConfiguration config)
+        {
+            var tokenHandler = new JwtSecurityTokenHandler();
+            var key = Encoding.ASCII.GetBytes(config["Jwt:Key"]);
+            var userIdClaim = new Claim("userId", user.UserId.ToString());
+            var tokenDescriptor = new SecurityTokenDescriptor
+            {
+                Subject = new ClaimsIdentity(new Claim[]
+                {
+                    new Claim(ClaimTypes.Email,user.Email),
+                    new Claim(ClaimTypes.Name,user.Name),
+                    userIdClaim
+                }),
+                Expires = DateTime.UtcNow.AddHours(1),
+                SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(key), SecurityAlgorithms.HmacSha256Signature)
+            };
+            var token = tokenHandler.CreateToken(tokenDescriptor);
+            var jwt = tokenHandler.WriteToken(token);
+
+            return new JsonResult(new { token = jwt });
+        }
+    }
+}
